Update organizations.md
Typo... the beginning of the sentence should read An organization...

Each user can create or be a member of an unlimited number of organizations. An organization is automatically created upon user registration, and the user is assigned as its administrator. This organization is named after the user's balenaCloud username and is managed like any other but currently cannot be deleted.